fix: use cookies instead of localstorage
